clase_10_JWT/requirements.txt
# app
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.1

# security
pyjwt==2.10.0
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
python-jose==3.5.0
cryptography==45.0.3